add translationArrayHelper
